requirements.txt
streamlit
easyocr
pillow
opencv-python
numpy
requests
openai
